requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
awscli==1.16.294
boto3==1.10.30
botocore==1.13.30
cassandra-driver==3.20.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
dask==2.8.1
defusedxml==0.6.0
Django==3.0
django-allauth==0.40.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-environ==0.4.5
django-extensions==2.2.5
django-model-utils==3.2.0
djangorestframework==3.10.3
docutils==0.15.2
factory-boy==2.12.0
Faker==3.0.0
fastparquet==0.3.2
fsspec==0.6.1
idna==2.8
importlib-metadata==1.1.0
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
kafka-python==1.4.7
llvmlite==0.30.0
locket==0.2.0
lockfile==0.12.2
luigi==2.8.10
MarkupSafe==1.1.1
more-itertools==8.0.0
numba==0.46.0
numpy==1.17.4
oauthlib==3.1.0
packaging==19.2
pandas==0.25.3
partd==1.1.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.0
py4j==0.10.7
pyasn1==0.4.8
pycparser==2.19
pyparsing==2.4.5
pyspark==2.4.4
pytest==5.3.1
pytest-cov==2.8.1
pytest-django==3.7.0
python-daemon==2.1.2
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3fs==0.4.0
s3transfer==0.2.1
setuptools-scm==3.3.3
six==1.13.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
text-unidecode==1.3
thrift==0.13.0
toolz==0.10.0
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
